Consolidate MealsMixin into Fatsecret class
